+++ components/applications/patina_applications_squashfs-tools.sh
@@ -24,22 +24,21 @@
 # Functions #
 #############
 
-# PATINA > FUNCTIONS > APPLICATIONS > GIT
+# PATINA > FUNCTIONS > APPLICATIONS > SQUASHFS-TOOLS
 
-patina_git() {
+patina_squashfs-tools() {
   # Success: Display help and exit.
   if [ "$1" = '--help' ] ; then
-    echo "Usage: p-git [OPTION]"
-    echo "Perform git operations within the current directory."
-    echo "Dependencies: 'git' command from package 'git'."
+    echo "Usage: p-squash [DIRECTORY] [OPTION]"
+    echo "Create a read-only disk image in SquashFS format."
+    echo "Dependencies: 'mksquashfs' command from package 'squashfs-tools'."
     echo
-    echo -e "  undo\\t\\tDiscard changes to repository since previous commit."
     echo -e "  --help\\tDisplay this help and exit."
     echo
     return 0
 
-  # Failure: Patina cannot detect a required application.
-  elif ( ! command -v 'git' > /dev/null 2>&1 ) ; then
+  # Failure: Command 'mksquashfs' is not available.
+  elif ( ! command -v 'mksquashfs' > /dev/null 2>&1 ) ; then
     patina_raise_exception 'PE0006'
     return 127
 
@@ -53,24 +52,47 @@ patina_git() {
     patina_raise_exception 'PE0002'
     return 1
 
-  # Success: Parse 'git' arguments.
-  elif [ "$1" = 'undo' ] ; then
-    git reset --hard HEAD
+  # Failure: The target exists, but is a file.
+  elif [ -f "$1" ] ; then
+    patina_raise_exception 'PE0014'
+    return 1
+
+  # Failure: The target directory does not exist.
+  elif [ ! -d "$1" ] ; then
+    patina_raise_exception 'PE0004'
+    return 1
+
+  # Failure: The target disk image already exists.
+  elif [ -f "$(basename "$1").sqsh" ] ; then
+    patina_raise_exception 'PE0011'
+    return 1
+
+  # Success: Create SquashFS disk image.
+  elif [ -d "$1" ] ; then
+    mksquashfs "$1" "$1.sqsh"
     return 0
 
-  # Failure: An invalid argument was provided.
+  # Failure: Catch all.
   else
-    patina_raise_exception 'PE0003'
+    patina_raise_exception 'PE0000'
     return 1
   fi
 }
 
 ###########
+# Exports #
+###########
+
+# PATINA > FUNCTIONS > APPLICATIONS > SQUASHFS-TOOLS
+
+export -f 'patina_squashfs-tools'
+
+###########
 # Aliases #
 ###########
 
-# PATINA > FUNCTIONS > APPLICATIONS > GIT COMMANDS
+# PATINA > FUNCTIONS > APPLICATIONS > SQUASHFS-TOOLS COMMANDS
 
-alias 'p-git'='patina_git'
+alias 'p-squash'='patina_squashfs-tools'
 
 # End of File.